Create a separate module for pyre's helper functions
