+++ libsa4py/utils.py
@@ -1,12 +1,11 @@
-from typing import List, Union
+from typing import List
 from tqdm import tqdm
 from joblib import Parallel
-from libcst.metadata.type_inference_provider import run_command, PyreData
 from os.path import join, isdir
 from pathlib import Path
-from datetime import datetime
 import time
 import os
+import signal
 import json
 
 
@@ -103,17 +102,3 @@ def find_repos_list(projects_path: str) -> List[dict]:
 def mk_dir_not_exist(path: str):
     if not isdir(path):
         os.mkdir(path)
-
-
-def pyre_server_init(project_path: str):
-    stdout, stderr, r_code = run_command("cd %s; echo -ne '.\n' | pyre init; pyre start" % project_path)
-
-
-def pyre_file_types(project_path: str, file_path: str) -> Union[PyreData, None]:
-    stdout, stderr, r_code = run_command('''cd %s; pyre query "types(path='%s')"''' % (project_path,
-                                                                               str(Path(file_path).relative_to(Path(project_path)))))
-    if r_code == 0:
-        return json.loads(stdout)["response"][0]
-    else:
-        print("PYRE_ERROR: ", stderr)
-        return None
